outsource `getStyleFromPayload` to `style` file
